docs: 添加 refactor 说明 refactor: 修改了Sidebar.vue的入口，使之规范化
